Create SECURITY.ND
First of all, if this coming from the Justice Department, Do you guys have Experts that can work unto getting this problem solved when it comes to Vulnerability? And in addition, your Justice Department here in ND sucks and it's all Fake charges on charges even with payments already made on time as usual, it's always a new stuff added to it. If it can be reviewed it will be amazing. Thank you 🙏